moving daal4py/sycl examples to tests
