+++ tests/daal4py/sycl/decision_forest_regression.py
@@ -1,5 +1,5 @@
 # ==============================================================================
-# Copyright 2021 Intel Corporation
+# Copyright 2020 Intel Corporation
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
 # limitations under the License.
 # ==============================================================================
 
-# daal4py Decision Forest Classification example of Hist method for shared memory systems
+# daal4py Decision Forest Regression example for shared memory systems
 
 import os
 
@@ -46,34 +46,23 @@ except Exception:
 
 
 # Commone code for both CPU and GPU computations
-def compute(train_data, train_labels, predict_data):
-    # Configure a training object (5 classes)
-    train_algo = d4p.decision_forest_classification_training(
-        5,
+def compute(train_data, train_labels, predict_data, method="defaultDense"):
+    # Configure a training object
+    train_algo = d4p.decision_forest_regression_training(
+        nTrees=100,
         fptype="float",
-        method="hist",
-        maxBins=256,
-        minBinSize=1,
-        nTrees=10,
-        minObservationsInLeafNode=8,
-        featuresPerNode=3,
-        engine=d4p.engines_mt19937(seed=777),
-        varImportance="MDI",
+        engine=d4p.engines_mt2203(seed=777),
+        varImportance="MDA_Raw",
         bootstrap=True,
-        resultsToCompute="computeOutOfBagError",
+        resultsToCompute="computeOutOfBagError|computeOutOfBagErrorPerObservation",
+        method=method,
     )
-
     # Training result provides (depending on parameters) model,
     # outOfBagError, outOfBagErrorPerObservation and/or variableImportance
     train_result = train_algo.compute(train_data, train_labels)
 
     # now predict using the model from the training above
-    predict_algo = d4p.decision_forest_classification_prediction(
-        nClasses=5,
-        fptype="float",
-        resultsToEvaluate="computeClassLabels|computeClassProbabilities",
-        votingMethod="unweighted",
-    )
+    predict_algo = d4p.decision_forest_regression_prediction(fptype="float")
 
     predict_result = predict_algo.compute(predict_data, train_result.model)
 
@@ -99,11 +88,11 @@ def to_numpy(data):
     return data
 
 
-def main(readcsv=read_csv):
-    nFeatures = 3
+def main(readcsv=read_csv, method="defaultDense"):
+    nFeatures = 13
     # input data file
-    train_file = os.path.join("..", "data", "batch", "df_classification_train.csv")
-    predict_file = os.path.join("..", "data", "batch", "df_classification_test.csv")
+    train_file = os.path.join("..", "..", "..", "examples", "daal4py", "data", "batch", "df_regression_train.csv")
+    predict_file = os.path.join("..", "..", "..", "examples", "daal4py", "data", "batch", "df_regression_test.csv")
 
     # Read train data. Let's use 3 features per observation
     train_data = readcsv(train_file, range(nFeatures), t=np.float32)
@@ -113,9 +102,11 @@ def main(readcsv=read_csv):
     predict_labels = readcsv(predict_file, range(nFeatures, nFeatures + 1), t=np.float32)
 
     # Using of the classic way (computations on CPU)
-    train_result, predict_result = compute(train_data, train_labels, predict_data)
+    train_result, predict_result = compute(
+        train_data, train_labels, predict_data, "defaultDense"
+    )
     assert predict_result.prediction.shape == (predict_labels.shape[0], 1)
-    assert (np.mean(predict_result.prediction != predict_labels) < 0.04).any()
+    assert (np.square(predict_result.prediction - predict_labels).mean() < 18).any()
 
     train_data = to_numpy(train_data)
     train_labels = to_numpy(train_labels)
@@ -128,10 +119,12 @@ def main(readcsv=read_csv):
             sycl_train_labels = sycl_buffer(train_labels)
             sycl_predict_data = sycl_buffer(predict_data)
             train_result, predict_result = compute(
-                sycl_train_data, sycl_train_labels, sycl_predict_data
+                sycl_train_data, sycl_train_labels, sycl_predict_data, "hist"
             )
             assert predict_result.prediction.shape == (predict_labels.shape[0], 1)
-            assert (np.mean(predict_result.prediction != predict_labels) < 0.03).any()
+            assert (
+                np.square(predict_result.prediction - predict_labels).mean() < 18
+            ).any()
 
     return (train_result, predict_result, predict_labels)
 
@@ -143,10 +136,6 @@ if __name__ == "__main__":
     print(
         "\nDecision forest prediction results (first 10 rows):\n",
         predict_result.prediction[0:10],
-    )
-    print(
-        "\nDecision forest probabilities results (first 10 rows):\n",
-        predict_result.probabilities[0:10],
     )
     print("\nGround truth (first 10 rows):\n", plabels[0:10])
     print("All looks good!")
