+++ tests/daal4py/sycl/linear_regression.py
@@ -79,12 +79,12 @@ def to_numpy(data):
 def main(readcsv=read_csv, method="defaultDense"):
     # read training data. Let's have 10 independent,
     # and 2 dependent variables (for each observation)
-    trainfile = os.path.join("..", "data", "batch", "linear_regression_train.csv")
+    trainfile = os.path.join("..", "..", "..", "examples", "daal4py", "data", "batch", "linear_regression_train.csv")
     train_indep_data = readcsv(trainfile, range(10), t=np.float32)
     train_dep_data = readcsv(trainfile, range(10, 12), t=np.float32)
 
     # read testing data
-    testfile = os.path.join("..", "data", "batch", "linear_regression_test.csv")
+    testfile = os.path.join("..", "..", "..", "examples", "daal4py", "data", "batch", "linear_regression_test.csv")
     test_indep_data = readcsv(testfile, range(10), t=np.float32)
     test_dep_data = readcsv(testfile, range(10, 12), t=np.float32)
