+++ tests/daal4py/sycl/sklearn_sycl.py
